api/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
stripe==14.4.1
resend==2.26.0
passlib[bcrypt]==1.7.4
pyjwt==2.12.1
slowapi==0.1.9
psycopg2-binary==2.9.11
python-multipart==0.0.22
python-jose[cryptography]==3.5.0
python-dotenv==1.2.2
